refactor: extract OBJECTS_BLOCKED_FROM_AUTOMATION into its own constants file (one export per file)
